Fix for #151. Two views for form
